Expand MessageWindow to include InputField and clear chat button; Moved generateSessionId into its own file under utils
